% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCNVRs.R
\name{cnvr_fast}
\alias{cnvr_fast}
\title{Compute Copy Number Variable Regions (CNVRs)}
\usage{
cnvr_fast(put_cnvs)
}
\arguments{
\item{cnvs}{lorem ipsum}

\item{chr_arms}{lorem ipsum}

\item{final_merge}{lorem ipsum}
}
\value{
a \code{list} of two elements. The first element is a \code{data.table}
that contains the actual CNVR information, genomic location and frequency in
the cohort. The second element is the \code{CNVresults}
}
\description{
Faster CNVRs computation algorithm
}
\details{
It is based on pamk clustering algorithm. It is a much simpler process that
takes advantage of the fact that this package is meant to deal only with groups
of calls overlapping a specific locus.
}
