% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRDS.R
\name{extractRDS}
\alias{extractRDS}
\title{Extratc Raw Data in Specific Loci}
\usage{
extractRDS(loci, samples_file, rds_path)
}
\description{
\code{extractRDS()} extract the raw LRR and BAF values in the loci from
the intensity files.
}
