% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tabix.R
\name{get_region_tabix}
\alias{get_region_tabix}
\title{Load raw data from tabix indexed intensity file}
\usage{
get_region_tabix(chr, start, end, f_path, snppos = NA)
}
\description{
Load raw data from tabix indexed intensity file
}
