% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_chr_arms}
\alias{hg38_chr_arms}
\title{Start and end of all chromosome arms in the assembly hg38}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{chr}{chromosome, 1:24}
\item{arm_ID}{name of the arm, chromosome name and p/q}
\item{start}{first base pair position, 0 indexed}
\item{end}{last position}
...
}
}
\usage{
hg38_chr_arms
}
\description{
Start and end of all chromosome arms in the assembly hg38
}
\keyword{datasets}
