% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_start_end_centromeres}
\alias{hg38_start_end_centromeres}
\title{Start, end and centrosome position of all chromosomes in the assembly hg38}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{chr}{chromosome, 1:24}
\item{start}{first base pair position, 0 indexed}
\item{end}{last position}
\item{centrosome}{centrosome position}
...
}
}
\usage{
hg38_start_end_centromeres
}
\description{
Start, end and centrosome position of all chromosomes in the assembly hg38
}
\keyword{datasets}
