% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCtree.R
\name{qctree}
\alias{qctree}
\title{CNVs QC Filtering Tree}
\usage{
qctree(
  cnvs,
  cnvrs,
  qsdt,
  loci,
  maxLRRSD = 0.3,
  maxBAFdrift = 0.01,
  maxGCWF = 0.015,
  minGCWF = -0.015,
  commonCNVRsMinFreq = NA,
  st4minlogr1 = -0.35,
  st4maxlogr1 = 0.4,
  st4maxBAFcDEL = 0.03,
  st4maxBAFcDUP = 0.125,
  st4maxBAFbDEL = 0.075,
  st4maxBAFbDUP = NA,
  st5maxLRRSDlocus,
  st5maxBAFcDEL = 0.05,
  st5maxBAFcDUP = 0.15,
  st5maxBAFbDEL = 0.1,
  st5maxBAFbDUP = NA,
  clean_out = T
)
}
\description{
The main functions of the \code{QCtreeCNV} package
}
