% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qctree_pipeline.R
\name{qctree_pipeline}
\alias{qctree_pipeline}
\title{QCtree pipeline in a convenient and automated form}
\usage{
qctree_pipeline(
  loci,
  calls,
  pennqc,
  samples_list,
  rds_path,
  cnvrs = NA,
  hg_version = c("gh18", "hg19", "hg38")
)
}
\description{
All steps from the PennCNV raw data to the objects required by
\code{qctree()}
}
\details{
Please note that at the moment having the input in the required format
is still something the user needs to take care of. Only defaults values
for each step are supported at the moment (will change soon)
}
