% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVs.R
\name{saveCNVplots}
\alias{saveCNVplots}
\title{Save CNVs plots for Visual Inspection}
\usage{
saveCNVplots(cnvs, plots_path, loci, sample_list, reg_len = 2e+06)
}
\arguments{
\item{cnvs}{lorem ipsum}

\item{plots_path}{lorem ipsum}

\item{loci}{lorem ipsum}

\item{sample_list}{lorem ipsum}

\item{reg_len}{lorem ipsum}
}
\description{
Plot LRR BAF plots and save them to be used in visual inspection
}
