% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stitch_filter_calls.R
\name{select_stitch_calls}
\alias{select_stitch_calls}
\title{Select calls in loci and stich close calls}
\usage{
select_stitch_calls(cnvs, loci, minsnp = 20, maxgap = 0.5, minoverlap = 0.2)
}
\arguments{
\item{cnvs}{lorem ipsum}

\item{loci}{lorem ipsum}

\item{minsnp}{lorem ipsum}

\item{maxgap}{lorem ipsum}

\item{minoverlap}{lorem ipsum}
}
\description{
Select calls in loci and stich close calls
}
