% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetrics.R
\name{extractMetrics}
\alias{extractMetrics}
\title{Extratc QC metrics}
\usage{
extractMetrics(loci, cnvs, pennQC, samples_list, snppos = NA)
}
\arguments{
\item{loci}{lorem ipsum}

\item{cnvs}{lorem ipsum}

\item{pennQC}{lorem ipsum}

\item{samples_list}{lorem ipsum}

\item{snppos}{lorem ipsum}
}
\description{
\code{extractMetrics()} extract the QC metrics from the raw data in RDS
format.
}
