% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetrics.R
\name{extractMetrics}
\alias{extractMetrics}
\title{Extratc QC metrics}
\usage{
extractMetrics(loci, cnvs, int_rds_path, tmp_rds_path)
}
\description{
\code{extractMetrics()} extract the QC metrics from the raw data in RDS
format.
}
