% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{qc_plots_cnvs}
\alias{qc_plots_cnvs}
\title{Create Quality Control plots for the filtering pipeline}
\usage{
qc_plots_cnvs(
  cnvs,
  folder_name,
  qc,
  maxLRRSD = 0.35,
  maxBAFdrift = 0.01,
  maxGCWF = 0.02,
  minGCWF = -0.02
)
}
\description{
Create Quality Control plots for the filtering pipeline
}
