% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCtree.R
\name{qctree}
\alias{qctree}
\title{CNVs QC Filtering Tree}
\usage{
qctree(
  cnvs,
  cnvrs,
  qsdt,
  loci,
  maxLRRSD = 0.35,
  maxBAFdrift = 0.01,
  maxGCWF = 0.02,
  minGCWF = -0.02,
  commonCNVRsMinFreq = NA,
  st2minov = 0.65,
  st4minlogr1 = -0.35,
  st4maxlogr1 = 0.4,
  st4maxBAFcDEL = 0.03,
  st4maxBAFcDUP = 0.125,
  st4maxBAFbDEL = 0.075,
  st5maxLRRSDlocus = 0.35,
  st5maxlogr1 = 0.5,
  st5maxBAFcDEL = 0.05,
  st5maxBAFcDUP = 0.15,
  st5maxBAFbDEL = 0.1,
  st5maxBAFbDUP = NA,
  clean_out = T,
  st5exclLRRSD = 0.65
)
}
\arguments{
\item{cnvs}{lorem ipsum}

\item{cnvrs}{lorem ipsum}

\item{qsdt}{lorem ipsum}

\item{loci}{lorem ipsum}

\item{maxLRRSD}{lorem ipsum}

\item{maxBAFdrift}{lorem ipsum}

\item{maxGCWF}{lorem ipsum}

\item{minGCWF}{lorem ipsum}

\item{commonCNVRsMinFreq}{lorem ipsum}

\item{st4minlogr1}{lorem ipsum}

\item{st4maxlogr1}{lorem ipsum}

\item{st4maxBAFcDEL}{lorem ipsum}

\item{st4maxBAFcDUP}{lorem ipsum}

\item{st4maxBAFbDEL}{lorem ipsum}

\item{st5maxLRRSDlocus}{lorem ipsum}

\item{st5maxBAFcDEL}{lorem ipsum}

\item{st5maxBAFcDUP}{lorem ipsum}

\item{st5maxBAFbDEL}{lorem ipsum}

\item{st5maxBAFbDUP}{lorem ipsum}

\item{clean_out}{lorem ipsum}

\item{st5exclLRRSD}{lorem ipsum}
}
\description{
The main functions of the \code{QCtreeCNV} package
}
